pyproject.toml
[tool.poetry]
name = "negrucoin"
version = "0.1.0"
description = "Fun site with passive money"
authors = ["IvanPragma <ivanms.ept@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.1"
djangorestframework = "^3.14.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.1"
django-cors-headers = "^3.14.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
pytest-django = "^4.5.2"
pylint = "^2.15.3"
pylint-django = "^2.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
